[feat] 添加AMP混合精度支持并优化模型评估
- 引入AMP自动混合精度训练，支持BF16精度
- 确保sigmoid操作在FP32下执行以保持精度
- 增加_get_amp_context工具函数统一管理AMP上下文
- 调整评估批处理大小以优化内存使用
- 优化EmbeddingModule的设备兼容性

+++ src/train_dft.py
@@ -335,7 +335,7 @@ def train_dft_model(model, train_stock_info, test_stock_info,
                 output = model(batch_inputs)
                 # 每次前向后重新计算DFT自引导权重（SAM第二步时参数已扰动，权重需更新）
                 with torch.no_grad():
-                    dft_w = compute_dft_weights(torch.sigmoid(output), w_min=dft_w_min, w_max=dft_w_max)
+                    dft_w = compute_dft_weights(torch.sigmoid(output.float()), w_min=dft_w_min, w_max=dft_w_max)
             # 根据损失函数类型选择计算方式
                 if use_task_aligned:
                 # 刷新正负样本动态权重（task_loss_fn内部BCE + 独立的dft_train_bce同步更新）
